refactor(storage/gcp): create client.go for initialization and registration
